Add temporary hit counter
It was brought to my attention that this post may get shared around a
bit. Attached a hit counter service provided by hits.sh temporarily in
order to get a rough idea of that.

In the future, every post should get a hit counter automatically via a
template or something, and ideally it would be hosted by the local
GitHub Actions rather than a third-party service (but there isn't time
to research this properly at the moment).