New post
An addendum - the first of likely many - to the Open Letter, covering
events since my ban. Specifcially, this focuses on the 3 month
suspension that Tim Peters has received from the Steering Council.

A few copyedits were made during the writing process - not saved as
edits this time - but unlike the original letter, this was basically
written in a single draft.