Initial setup for DPO posts collection
A folder is added with the Markdown source for the posts in question.
_config.yml is modified to reference the collection, and a dpo.md file
is added to the header to introduce the collection.

The posts still need to be edited to convert quote tags to Markdown
blockquotes and add front matter; and the introduction needs to be
edited with a custom summary of those pages and appropriate links.