Reorganize old posts
Post slugs / titles now no longer include the date, andd the posts are
in YMD folders (new ones will be automatically placed via the `new_post`
command).

This breaks old internal links, and a bunch of redirects also need to be
set up anyway because others may have bookmarks.

Additional work is still necessary on the category/tag setup.

Ideally, the post slug should be inferred from the title and that
inferred value should be used in the {{% hitcounter %}} template.