Don't use static indexes
The "static index" feature is an approach to pagination where old posts
don't change which page they're on once they leave the first page.
However, for this to work, the next most recent page after the front
page needs to have a varying number of posts (accumulating up to 5
before a new page is added to the pagination).

I'm not interested in having these static links to old index pages
because they're only meant as a display of the catalog; I do care about
the aesthetic of having a full page behind the front page, and leaving
any remaining fraction of posts (not divisible by the posts-per-page
count - currently 5) on the oldest page.