Rewrite of teaser
The original content for the opening has been removed, since part 0 of
the series isn't written yet and because I have thus far been assuming a
certain level of familiarity.

The new teaser includes a very brief explanation of the key concepts,
and is focused on describing The Bug without any sense of mystery.