Reworked "The Big Reveal"
The old bullet points were not as neatly organized as they appeared.
Added a prose description of what's really so ridiculous about Pip
trying to determine a package name and version this way, giving clearer
emphasis to the multiple ways this information is already available.

Also added/corrected a note about what happens when this information is
invalid (it doesn't appear to error out, in current testing). Perhaps
this deserves a separate section with an appropriately modified test
script.